Renaming for multiple types of scouting data
