+++ public/protobuf/data_transfer.proto
@@ -4,11 +4,9 @@ message DataTransfer {
     QrCodeType qrType = 1;
     string version = 2;
 
-    repeated MatchData matches = 3;
-    repeated MatchEventData events = 4;
-    repeated ScheduledMatch schedule = 5;
-    optional ScheduleData scheduleData = 6;
-    optional PickListData pickListData = 7;
+    optional ScoutData scoutData = 3;
+    optional ScheduleData scheduleData = 4;
+    optional PickListData pickListData = 5;
 }
 
 enum QrCodeType {
@@ -17,14 +15,12 @@ enum QrCodeType {
     PickList = 2;
 }
 
-message MatchEventData {
-    string matchId = 1;
-    int32 teamNumber = 2;
-    int32 event = 3;
-    int64 time = 4;
-}
+// Scouting Data
 
-// Match Data
+message ScoutData {
+    repeated MatchData matches = 1;
+    repeated MatchEventData events = 2;
+}
 
 message MatchData {
     // header
@@ -32,7 +28,7 @@ message MatchData {
     int32 teamNumber = 2;
     AllianceColor allianceColor = 3;
 
-    // fields
+    // fields ## EDIT THESE ##
     HumanPlayerLocation humanPlayerLocation = 4;
     bool preload = 5;
     ClimbResult climb = 6;
@@ -45,6 +41,13 @@ message MatchData {
     string scoutName = 11;
     int64 matchStart = 12;
     int64 submitTime = 13;
+}
+
+message MatchEventData {
+    string matchId = 1;
+    int32 teamNumber = 2;
+    int32 event = 3;
+    int64 time = 4;
 }
 
 enum AllianceColor {
@@ -72,6 +75,13 @@ enum MatchResult {
 
 // Schedule Data
 
+message ScheduleData {
+    repeated ScheduledMatch schedule = 1;
+    bool fieldRotated = 2;
+    string competitionId = 3;
+    int32 currentMatch = 4;
+}
+
 message ScheduledMatch {
     string matchId = 1;
     int32 blue1 = 2;
@@ -82,15 +92,10 @@ message ScheduledMatch {
     int32 red3 = 7;
 }
 
-message ScheduleData {
-    bool fieldRotated = 1;
-    string competitionId = 2;
-    int32 currentMatch = 3;
-}
+// Pick List Data
 
 message PickListData {
     repeated int32 pickList = 1;
     repeated int32 crossedOut = 2;
-
     string competitionId = 3;
 }
