+++ src/enums/QrCodeType.ts
@@ -2,7 +2,7 @@
  * Enum for 
  */
 enum QrCodeType {
-    MatchData = 0,
+    ScoutData = 0,
     Schedule = 1,
     PickList = 2,
 }
